Add the formating/impsort plugin
This will automatically format the build according to the
eclipse code style provided in etc/config. In order to
validate that the build is correctly formatted CI is run
with -Dvalidate-format which will verify that everything
is correct.

This plugin uses the Eclipse code formatter, which can also
be embedded into Intellij via a plugin. This means that
everyone shuold be able to use the exact same formatter and
get a 100% consistent format with no effort required.

Signed-off-by: Stuart Douglas <sdouglas@redhat.com>