Add implementations for Cookie.equals, Cookie.toString, and Cookie.hashcode
This is good practice for any "real" class, and there are some practical reasons specifically for `equals` and `toString`.

When writing a unit-test for some servlet-related code, I was surprised to find out that two Cookie objects were not `equals()` each other even though their fields were identical. Of course, I first had to subclass `Cookie` and then override `toString` to find out that they actually *were* equal, and then discovered that `Cookie.equals` hadn't been implemented, relying on the implementation of `Object.equals` which performs a reference-equality check *only*. Only after I changed my code over to use this Cookie subclass was I able to demonstrate via unit-test that my code was working.

This was *possible* in my case because I was controlling both the input and output Cookie objects. But in some other case, one might expect the servlet container to create the new Cookie object(s) and be unable to write a (sane) test that passed due to these missing methods.

I therefore request the inclusion of these methods into the Cookie class.