Trivial change to fix copyright check error.
The copyright check looks at last authored date. The merge of #499
retained the original author date of 2020 for the change to this file.
The date of 2023 in the file is inconsistent with that so the copyright
check fails. The sole purpose of this commit is to changed the authored
date of the most recent change to a date in 2023.